ui: move Header component to home folder
