+++ src/frontend/app/globals.css
@@ -68,3 +68,16 @@
         --chart-5: 340 75% 55%;
     }
 }
+
+
+
+
+
+@layer base {
+  * {
+    @apply border-border outline-ring/50;
+    }
+  body {
+    @apply bg-background text-foreground;
+    }
+}
